ci(release): Add automated release action
To issue a new release, we just need to bump the release
version in the `VERSION` file, populate `CHANGELOG.md`
with the Release notes and we're good to go. This patch
adds the `release-trigger.yaml` workflow which automatically
tags the repo, runs the CI, creates the github release and
uploads the relevant binaries as assets.

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>