chore(docs): Update Linux tutorial to note that urunit is needed in K8s
When we deploy a Linux container with urunc in Kubernetes, we face the
issue where the gateway might be in a different subnet than the IP.
In such cases, Linux will fail to set the default gateway and as such,
the VM will not have network access. Since urunc uses kernel boot
parameters to set up the network configuration of the VM, there is no
way to properly handle such issues.

As a result, we can only overcome this by an init process running inside
the VM. That is what urunit does, by setting the default route through
eth0. However, this makes urunit necessary when we deploy
such containers in K8s. Therefore, we need to mention this in our tutorial.

With this opportunity remove the rumprun debug message for the JSON
construction.

PR: https://github.com/urunc-dev/urunc/pull/227
Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>