fix(docs): Replace old image names with new ones
Replace the image reference in some exmaple unikernels in the docs with
the updated image reference that also reflects the need to mount
container's rootfs as the VM's rootfs. This is implied by the keyword
"raw" in the end of the image reference.

Also these images make use of the new mountRootfs annotation, which
replaced the obsolete useDMBlock one.

PR: https://github.com/urunc-dev/urunc/pull/231
Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>